+++ srcs/execution/heredoc.c
@@ -6,7 +6,7 @@
 /*   By: mstegema <mstegema@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/07/04 14:59:07 by cschabra      #+#    #+#                 */
-/*   Updated: 2023/08/31 14:54:52 by cschabra      ########   odam.nl         */
+/*   Updated: 2023/08/31 17:12:23 by cschabra      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ static void	ft_expand(void)
 static void	ft_read_input(char *data, int datalen, int *fd, bool expand)
 {
 	char	*temp;
-	int		templen;
+	size_t	templen;
 
 	while (1)
 	{
@@ -28,7 +28,7 @@ static void	ft_read_input(char *data, int datalen, int *fd, bool expand)
 		if (!temp)
 		{
 			perror("minishell :");
-			break ;
+			break ; // or free all and exit?
 		}
 		if (!ft_strncmp(temp, data, datalen))
 		{
@@ -68,14 +68,14 @@ static void	ft_remove_quotes(char *data)
 void	ft_heredoc(char *data)
 {
 	int		fd[2];
-	int		datalen;
+	size_t	datalen;
 	bool	expand;
 
 	expand = true;
 	if (pipe(fd) == -1)
 	{
 		perror("BabyBash");
-		return ;
+		return ; // free all and exit???
 	}
 	datalen = ft_strlen(data);
 	if (ft_strchr(data, '"') || ft_strchr(data, '\''))
